Align josh init tsconfig extends order with sync (kit preset first) #561
