+++ package.json
@@ -1,5 +1,5 @@
 {
-	"version": "0.254.0",
+	"version": "0.255.0",
 	"name": "@joshuafolkken/kit",
 	"type": "module",
 	"license": "MIT",
